chore: refactor motis helpers into their own folder
