+++ visual-debugger/src/lib/components/ui/Route.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-    import {getModeStyle, routeColor, type LegLike} from '@/modeStyle.ts';
-    import {cn} from '@/utils.ts';
+    import {getModeStyle, routeColor, type LegLike} from '@/motis-helpers/modeStyle.ts';
+    import {cn} from '@/motis-helpers/utils.ts';
     import type {LegShadow} from "@data/type-declarations/shadowTypes.ts";
     import MatchIndicator from "@/components/ui/subcomponents/MatchIndicator.svelte";
