+++ visual-debugger/src/lib/components/ui/card/card-header.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { HTMLAttributes } from "svelte/elements";
-	import { cn } from "$lib/utils.js";
+	import { cn } from "@/motis-helpers/utils.js";
 
 	type $$Props = HTMLAttributes<HTMLDivElement>;
