+++ visual-debugger/src/lib/components/ui/checkbox/checkbox.svelte
@@ -2,7 +2,7 @@
 	import { Checkbox as CheckboxPrimitive } from "bits-ui";
 	import Check from "lucide-svelte/icons/check";
 	import Minus from "lucide-svelte/icons/minus";
-	import { cn } from "$lib/utils.js";
+	import { cn } from "@/motis-helpers/utils.js";
 
 	type $$Props = CheckboxPrimitive.Props;
 	type $$Events = CheckboxPrimitive.Events;
