+++ visual-debugger/src/lib/components/ui/input/input.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import type { HTMLInputAttributes } from "svelte/elements";
 	import type { InputEvents } from "./index.js";
-	import { cn } from "$lib/utils.js";
+	import { cn } from "@/motis-helpers/utils.js";
 
 	type $$Props = HTMLInputAttributes;
 	type $$Events = InputEvents;
