+++ visual-debugger/src/lib/components/ui/radio-group/radio-group.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { RadioGroup as RadioGroupPrimitive } from "bits-ui";
-	import { cn } from "$lib/utils.js";
+	import { cn } from "@/motis-helpers/utils.js";
 
 	type $$Props = RadioGroupPrimitive.Props;
