+++ visual-debugger/src/lib/components/ui/scroll-area/scroll-area-scrollbar.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { ScrollArea as ScrollAreaPrimitive } from "bits-ui";
-	import { cn } from "$lib/utils.js";
+	import { cn } from "@/motis-helpers/utils.js";
 
 	type $$Props = ScrollAreaPrimitive.ScrollbarProps & {
 		orientation?: "vertical" | "horizontal";
