+++ visual-debugger/src/lib/components/ui/scroll-area/scroll-area.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { ScrollArea as ScrollAreaPrimitive } from "bits-ui";
 	import { Scrollbar } from "./index.js";
-	import { cn } from "$lib/utils.js";
+	import { cn } from "@/motis-helpers/utils.js";
 
 	type $$Props = ScrollAreaPrimitive.Props & {
 		orientation?: "vertical" | "horizontal" | "both";
