+++ visual-debugger/src/lib/components/ui/separator/separator.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { Separator as SeparatorPrimitive } from "bits-ui";
-	import { cn } from "$lib/utils.js";
+	import { cn } from "@/motis-helpers/utils.js";
 
 	type $$Props = SeparatorPrimitive.Props;
