+++ visual-debugger/src/lib/components/ui/table/table-cell.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { HTMLTdAttributes } from "svelte/elements";
-	import { cn } from "$lib/utils.js";
+	import { cn } from "@/motis-helpers/utils.js";
 
 	type $$Props = HTMLTdAttributes;
