+++ visual-debugger/src/lib/components/ui/toggle/toggle.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { Toggle as TogglePrimitive } from "bits-ui";
 	import { type Size, type Variant, toggleVariants } from "./index.js";
-	import { cn } from "$lib/utils.js";
+	import { cn } from "@/motis-helpers/utils.js";
 
 	type $$Props = TogglePrimitive.Props & {
 		variant?: Variant;
