+++ visual-debugger/src/lib/motis-helpers/formatDuration.ts
