+++ visual-debugger/src/lib/motis-helpers/i18n/translation.ts
@@ -1,4 +1,4 @@
-import en from './en';
+import en from './en.ts';
 
 export type Translations = {
 	journeyDetails: string;
