+++ visual-debugger/src/lib/motis-helpers/utils.ts
